[TASK 1] Refactor tests and docs, update CI
