+++ project/__init__.py
@@ -1,1 +1,2 @@
-print("import sources directory")
+import project.utils
+from project.utils import graph_utils
